refactor(cli): extract sandbox live state helpers
